Fix dtype array in stats

+++ file_comparison/report_generator.py
@@ -8,7 +8,7 @@ error_diff_types = ["type", "len"]
 
 
 def compute_1el_difference (origin, new):
-    return compute_1list_difference (np.array(origin, ndmin=1, dtype=float), np.array(new, ndmin=1, dtype=float))
+    return compute_1list_difference (np.array(origin, ndmin=1), np.array(new, ndmin=1))
 
 
 def compute_1list_difference (origin:np.ndarray, new:np.ndarray):
